[C++][DCGAN] Increment batch_index before the if statements?
I ran the example codes for [C++][DCGAN] but the 2 if statements in the nested for loop are not behaving as expected. The first one just prints the losses every epoch essentially since batch_index gets value 1 through the life of the program. As for the second one it seems to not save anything for the same reason as the first. Not sure if this code runs perfectly fine on other versions as I ran it on C++14. If I'm in the wrong then sorry I wasted your time with this PR.

Thanks