Fix load model when a single GPU is used
This change is necessary for --resume and --evaluate methods when --gpu is specified. For example, if the model was trained on GPU 3 and then you try to evaluate it on a machine with a single GPU called GPU 0, you would get this error: 

RuntimeError: Attempting to deserialize object on CUDA device 3 but torch.cuda.device_count() is 3. Please use torch.load with map_location to map your storages to an existing device.

This change allows --evaluate and --resume functionalities to work correctly when a single GPU is used.