Adding Distributed Model Parallel Training Example for RNN
The RNNModel puts the embedding table and the decoder parameters on
a remote parameter server, and locally holds parameters for the LSTM
module. The RNNModel overrides the forward() function to pass input
and intermediate outputs across workers. And it also added a
parameter_rrefs to return a list of RRefs of the model parameters.